Extract FeatureBlock Title and P
